[Refactor] 검색어 입력창 컴포넌트 분리
- 검색어 입력창 컴포넌트 개별 컴포넌트로 분리
